+++ lib/src/flutter_deck_slide.dart
@@ -60,9 +60,11 @@ class _SlideBody extends StatelessWidget {
     final scaffoldState = Scaffold.of(context);
 
     return FlutterDeckDrawerListener(
-      onDrawerToggle: scaffoldState.isDrawerOpen
-          ? scaffoldState.closeDrawer
-          : scaffoldState.openDrawer,
+      onDrawerToggle: () {
+        scaffoldState.isDrawerOpen
+            ? scaffoldState.closeDrawer()
+            : scaffoldState.openDrawer();
+      },
       child: child,
     );
   }